Moved User schema to schema folder
