+++ backend/schemas/User.js
@@ -1,0 +1,26 @@
+import mongoose from "mongoose";
+import crypto from "crypto";
+
+const UserSchema = new mongoose.Schema({
+	username: {
+		type: String,
+		required: true,
+		unique: true,
+	},
+	email: {
+		type: String,
+		unique: true
+	},
+	password: {
+		type: String,
+		required: true,
+	},
+	accessToken: {
+		type: String,
+		default: () => crypto.randomBytes(128).toString("hex"),
+	},
+});
+
+const User = mongoose.model("User", UserSchema);
+
+export default User;
